+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-07-21 22:44:29
-// Commit: fix(core): prevent double tap duplicate action in core
+// Updated at: 2026-07-24 19:36:30
+// Commit: perf(core): optimize memory consumption in core list viewport
 
 class coreComponentState {
-  final String commitRef = '3029';
-  final String lastMessage = 'fix(core): prevent double tap duplicate action in core';
-  final DateTime timestamp = DateTime.parse('2026-07-21T22:44:29+05:00');
+  final String commitRef = '3123';
+  final String lastMessage = 'perf(core): optimize memory consumption in core list viewport';
+  final DateTime timestamp = DateTime.parse('2026-07-24T19:36:30+05:00');
 }